+++ packages/pre-build/src/utils.js
@@ -91,9 +91,9 @@ const tsconfigPath = path.resolve(
 );
 exports.tsconfigPath = tsconfigPath;
 
-function getCompiledPath(sourceFilePath, sourceType) {
+function getCompiledPath(sourceFilePath, sourceType, sourceFiles) {
   const relativeSourcePath = '.' + sourceFilePath.replace(cwd, '');
-  const type = sourceType.check(relativeSourcePath.replace(/^.\//, ''));
+  const type = sourceType.check(relativeSourcePath.replace(/^.\//, ''), sourceFiles);
   const interTargetPath = path.resolve(
     distDir,
     sourceFilePath.replace(exports.baseDir + path.sep, ''),
@@ -134,3 +134,20 @@ exports.getNPMCommand = async () => {
 
   throw new Error(`Can not find these commands: ${candidate.join(', ')}`);
 };
+
+exports.execCallback = async (filePath, onSuccess) => {
+  const realOnSuccess = onSuccess || process.env.ON_SUCCESS_CALLBACK;
+  if (!realOnSuccess) {
+    return;
+  }
+  try {
+    await exports.exec(realOnSuccess, {
+      color: true,
+      env: {
+        HANDLED_FILE_PATH: filePath,
+      },
+    });
+  } catch (e) {
+    error(e);
+  }
+}